+++ packages/commerce-sdk-react/src/provider.tsx
@@ -23,7 +23,12 @@ import {
 import Auth from './auth'
 import {ApiClientConfigParams, ApiClients} from './hooks/types'
 import {Logger} from './types'
-import {MOBIFY_PATH, SLAS_PRIVATE_PROXY_PATH} from './constant'
+import {
+    DWSID_COOKIE_NAME,
+    MOBIFY_PATH,
+    SERVER_AFFINITY_HEADER_KEY,
+    SLAS_PRIVATE_PROXY_PATH
+} from './constant'
 export interface CommerceApiProviderProps extends ApiClientConfigParams {
     children: React.ReactNode
     proxy: string
@@ -33,12 +38,13 @@ export interface CommerceApiProviderProps extends ApiClientConfigParams {
     fetchOptions?: ShopperBasketsTypes.FetchOptions
     headers?: Record<string, string>
     fetchedToken?: string
-    OCAPISessionsURL?: string
     enablePWAKitPrivateClient?: boolean
     clientSecret?: string
     silenceWarnings?: boolean
     logger?: Logger
     defaultDnt?: boolean
+    refreshTokenRegisteredCookieTTL?: number
+    refreshTokenGuestCookieTTL?: number
 }
 
 /**
@@ -112,20 +118,65 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
         locale,
         currency,
         fetchedToken,
-        OCAPISessionsURL,
         enablePWAKitPrivateClient,
         clientSecret,
         silenceWarnings,
         logger,
-        defaultDnt
+        defaultDnt,
+        refreshTokenRegisteredCookieTTL,
+        refreshTokenGuestCookieTTL
     } = props
 
     // Set the logger based on provided configuration, or default to the console object if no logger is provided
     const configLogger = logger || console
 
+    const auth = useMemo(() => {
+        return new Auth({
+            clientId,
+            organizationId,
+            shortCode,
+            siteId,
+            proxy,
+            redirectURI,
+            fetchOptions,
+            fetchedToken,
+            enablePWAKitPrivateClient,
+            clientSecret,
+            silenceWarnings,
+            logger: configLogger,
+            defaultDnt,
+            refreshTokenRegisteredCookieTTL,
+            refreshTokenGuestCookieTTL
+        })
+    }, [
+        clientId,
+        organizationId,
+        shortCode,
+        siteId,
+        proxy,
+        redirectURI,
+        fetchOptions,
+        fetchedToken,
+        enablePWAKitPrivateClient,
+        clientSecret,
+        silenceWarnings,
+        configLogger,
+        refreshTokenRegisteredCookieTTL,
+        refreshTokenGuestCookieTTL
+    ])
+
+    const dwsid = auth.get(DWSID_COOKIE_NAME)
+    const serverAffinityHeader: Record<string, string> = {}
+    if (dwsid) {
+        serverAffinityHeader[SERVER_AFFINITY_HEADER_KEY] = dwsid
+    }
+
     const config = {
         proxy,
-        headers,
+        headers: {
+            ...headers,
+            ...serverAffinityHeader
+        },
         parameters: {
             clientId,
             organizationId,
@@ -171,39 +222,6 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
         headers?.['correlation-id']
     ])
 
-    const auth = useMemo(() => {
-        return new Auth({
-            clientId,
-            organizationId,
-            shortCode,
-            siteId,
-            proxy,
-            redirectURI,
-            fetchOptions,
-            fetchedToken,
-            OCAPISessionsURL,
-            enablePWAKitPrivateClient,
-            clientSecret,
-            silenceWarnings,
-            logger: configLogger,
-            defaultDnt
-        })
-    }, [
-        clientId,
-        organizationId,
-        shortCode,
-        siteId,
-        proxy,
-        redirectURI,
-        fetchOptions,
-        fetchedToken,
-        OCAPISessionsURL,
-        enablePWAKitPrivateClient,
-        clientSecret,
-        silenceWarnings,
-        configLogger
-    ])
-
     // Initialize the session
     useEffect(() => void auth.ready(), [auth])
 
@@ -222,7 +240,9 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
                 currency,
                 silenceWarnings,
                 logger: configLogger,
-                defaultDnt
+                defaultDnt,
+                refreshTokenRegisteredCookieTTL,
+                refreshTokenGuestCookieTTL
             }}
         >
             <CommerceApiContext.Provider value={apiClients}>