+++ packages/template-retail-react-app/app/components/product-view/index.jsx
@@ -594,6 +594,7 @@ const ProductView = forwardRef(
                                             // component, potentially as a prop called `selectInputOnFocus`.
                                             e.target.select()
                                         }}
+                                        productName={product?.name}
                                     />
                                 </VStack>
                             )}